scripts/requirements.txt
progressbar2
numpy
cv2